Add files via upload
In this project we have designed a packet filtering firewall.
Which works (Accept and reject the Packet ) based on the given rules.
It is written in python. 
We have used the csv library.
The csv module implements classes to read and write tabular data   in CSV format. . It presents classes and methods to perform read/write operations on CSV file.  A CSV file stores tabular data (numbers and text) in plain text. Each line of the file is a data record.

The rule is written in csv file so that ‘s why we have included this module.
Basically it will open the rule file and will take packet from interface_1.txt and it will match with the rule. If it matches, then it will reject the packet otherwise it will accept the packet.